pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "webhook2mqtt"
version = "0.1.0"
description = ""
authors = [
    {name = "Emil Djupvik",email = "emil@djupvik.dev"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "paho-mqtt (>=2.1.0,<3.0.0)",
    "colorlog (>=6.9.0,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
